Cambios urgentes muy urgentes
Correcciones de vaccine, vaccinationschedule para los tests

+++ src/main/java/ar/edu/unlp/info/bd2/model/Vaccine.java
@@ -1,8 +1,6 @@
 package ar.edu.unlp.info.bd2.model;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 @Entity
 public class Vaccine {
@@ -17,7 +15,7 @@ public class Vaccine {
     }
 
     @Column(name="ID")
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
